fix warnings

+++ src/client/CClient_Parse.cpp
@@ -211,7 +211,7 @@ void CClientNetEngine::ParseChallenge(CBytestream *bs)
 	bytestr.writeInt(PROTOCOL_VERSION,1);
 	bytestr.writeInt(client->iChallenge,4);
 	bytestr.writeInt(client->iNetSpeed,1);
-	bytestr.writeInt(client->connectInfo->worms.size(), 1);
+	bytestr.writeInt((int)client->connectInfo->worms.size(), 1);
 
 	// Send my worms info
     //
